Moved command files
